backend/requirements.txt
fastapi
uvicorn
pillow
numpy
scipy
pydub
cryptography
python-multipart
sqlalchemy
aiofiles
pytest
audioop-lts
